pyproject.toml
[tool.poetry]
name = "Squall"
version = "0.1.0"
description = "Security Monitoring and Scanning Container for Kubernetes"
authors = ["Bradley Sacks <bradsacks99@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
python-nmap = "^0.6.1"
fastapi = "^0.61.1"
uvicorn = "^0.12.2"
APScheduler = "^3.6.3"
SQLAlchemy = "1.3.22"
databases = "^0.4.1"
aiosqlite = "^0.16.0"
pysyncobj = "^0.3.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-pylint = "^0.18.0"
pytest-pep8 = "^1.0.6"
pytest-mypy = "^0.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
